requirements.txt
Django>=5.0.4,<5.1.0
black==24.4.2
djangorestframework==3.15.1
psycopg2== 2.9.9
pillow==10.3.0
python-dotenv==1.0.1
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1